// src/app/shared/services/api/model/get-all-rental-response.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetAllRentalResponse { 
    id?: number;
    carId?: number;
    rentStartDate?: string;
    rentEndDate?: string;
    renterName?: string;
    isCancelled?: boolean;
    createdDate?: string;
}